src/main/java/com/ninlgde/jcip/chapter14/boundedbuf/BoundedBuffer.java
package com.ninlgde.jcip.chapter14.boundedbuf;

/**
 * @author: ninlgde
 * @date: 1/27/21 11:56 AM
 */
public class BoundedBuffer<V> extends BaseBoundedBuffer<V> {

    protected BoundedBuffer(int capacity) {
        super(capacity);
    }

    public synchronized void put(V v) throws InterruptedException {
        while (isFull())
            wait();
        doPut(v);
        notifyAll();
    }

    public synchronized V take() throws InterruptedException {
        while (isEmpty())
            wait();
        V v = doTake();
        notifyAll();
        return v;
    }
}
